+++ wgpu-types/src/lib.rs
@@ -1062,14 +1062,12 @@ impl Limits {
     #[must_use]
     pub const fn using_recommended_minimum_mesh_shader_values(self) -> Self {
         Self {
-            // Literally just made this up as 256^2 or 2^16.
-            // My GPU supports 2^22, and compute shaders don't have this kind of limit.
-            // This very likely is never a real limiter
-            max_task_workgroup_total_count: 65536,
-            max_task_workgroups_per_dimension: 256,
+            // This is a common limit for apple devices. It's not immediately clear why.
+            max_task_workgroup_total_count: 1024,
+            max_task_workgroups_per_dimension: 1024,
             // llvmpipe reports 0 multiview count, which just means no multiview is allowed
             max_mesh_multiview_view_count: 0,
-            // llvmpipe once again requires this to be 8. An RTX 3060 supports well over 1024.
+            // llvmpipe once again requires this to be <=8. An RTX 3060 supports well over 1024.
             max_mesh_output_layers: 8,
             ..self
         }